+++ src/components/atomic-component/enchanted-textfield.ts
@@ -13,7 +13,7 @@
  * limitations under the License.                                           *
  * ======================================================================== */
 // External imports
-import { html, nothing } from 'lit';
+import { html, nothing, TemplateResult } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import { localized } from '@lit/localize';
 import { debounce } from 'lodash';
@@ -28,9 +28,6 @@ import { LOCALE_DIRECTIONS } from '../constants';
 import { EnchantedInputFieldType } from '../../types/enchanted-select';
 import { INPUT_TEXTFIELD_PARTS } from '../../types/cssClassEnums';
 import { AutoCompleteTextfieldEnum } from '../../types/enchanted-textfield';
-
-// Icon imports
-import './enchanted-svg-icon';
 import { KeyboardInputKeys } from '../../utils/keyboardEventKeys';
 
 const debug = createDebug('enchanted-web-components:components:ac:enchanted-textfield.ts');
@@ -61,11 +58,11 @@ export class EnchantedInputTextfield extends EnchantedAcBaseElement {
   @property({ type: Boolean })
   disabled = false;
 
-  @property({ type: String })
-  clearIconUrl = '';
+  @property()
+  clearIcon?: TemplateResult;
 
-  @property({ type: String })
-  actionIconUrl = '';
+  @property()
+  actionIcon?: TemplateResult;
 
   @property()
   field: EnchantedInputFieldType | string = '';
@@ -85,6 +82,14 @@ export class EnchantedInputTextfield extends EnchantedAcBaseElement {
   @state()
   private isRTL = getCurrentDirection() === LOCALE_DIRECTIONS.RTL;
 
+  private get hasClear(): boolean {
+    return !!this.clearIcon;
+  }
+
+  private get hasAction(): boolean {
+    return !!this.actionIcon;
+  }
+  
   connectedCallback(): void {
     super.connectedCallback();
   }
@@ -218,10 +223,10 @@ export class EnchantedInputTextfield extends EnchantedAcBaseElement {
       case INPUT_TEXTFIELD_PARTS.INPUT: {
         part = INPUT_TEXTFIELD_PARTS.INPUT;
         if (this.disabled) part = `${INPUT_TEXTFIELD_PARTS.INPUT} ${INPUT_TEXTFIELD_PARTS.INPUT_DISABLED}`;
-        if (this.clearIconUrl || this.actionIconUrl) part = `${part} ${this.isRTL
+        if (this.hasClear || this.hasAction) part = `${part} ${this.isRTL
           ? INPUT_TEXTFIELD_PARTS.INPUT_ICON_CLEAR_RTL
           : INPUT_TEXTFIELD_PARTS.INPUT_ICON_CLEAR}`;
-        if (this.clearIconUrl && this.actionIconUrl) part = `${part} ${this.isRTL
+        if (this.hasClear && this.hasAction) part = `${part} ${this.isRTL
           ? INPUT_TEXTFIELD_PARTS.INPUT_ICON_BOTH_RTL
           : INPUT_TEXTFIELD_PARTS.INPUT_ICON_BOTH}`;
       }
@@ -254,7 +259,7 @@ export class EnchantedInputTextfield extends EnchantedAcBaseElement {
     return html`
       <div part="div">
         ${this.label
-          ? html`<label data-testid="enchanted-input-textfield-label" for=${`input-${this.field}`} part="${this.getInputParts(INPUT_TEXTFIELD_PARTS.LABEL)}">${this.label}</label>`
+          ? html`<label data-testid="enchanted-textfield-label" for=${`input-${this.field}`} part="${this.getInputParts(INPUT_TEXTFIELD_PARTS.LABEL)}">${this.label}</label>`
           : nothing }
         <input
           tabIndex=1
@@ -274,34 +279,33 @@ export class EnchantedInputTextfield extends EnchantedAcBaseElement {
           aria-label=${this.ariaLabel || this.placeholder || this.getMessage('input.textfield.placeholder.type.to.search')}
         />
         <!-- This icon will take color from the parent component as useCurrentColor set to true -->
-        ${this.clearIconUrl
+        ${this.hasClear
           ? html`
-          <enchanted-svg-icon
+          <div
             tabIndex=2
             @click=${this.handleClear}
             @keydown=${this.handleClearEnter}
             data-testid="enchanted-clear-icon"
-            .icon=${this.clearIconUrl}
-            ?useCurrentColor=${true}
             aria-label=${this.getMessage('input.textfield.clear')}
             part="${this.getInputParts(INPUT_TEXTFIELD_PARTS.ICON_CLEAR)}"
             role="button"
-          />`
+          >
+            ${this.clearIcon }
+          </div>`
           : nothing}
-        <!-- This icon will take color from the parent component as useCurrentColor set to true -->
-        ${this.actionIconUrl
+        ${this.hasAction
           ? html`
-          <enchanted-svg-icon
-            .icon=${this.actionIconUrl}
-            ?useCurrentColor=${false}
+          <div
             @click=${this.handleSearch}
             @keydown=${this.handleEnterSearch}
-            tabIndex=${this.disabled ? -1 : 3}
             data-testid="enchanted-action-icon"
             aria-label=${this.getMessage('input.textfield.action')}
             part="${this.getInputParts(INPUT_TEXTFIELD_PARTS.ICON_ACTION)}"
             role="button"
-          />`
+            tabindex=${this.disabled ? -1 : 3}
+          >
+            ${this.actionIcon}
+          </div>`
           : nothing}
       </div>
     `;    
@@ -312,5 +316,4 @@ declare global {
   interface HTMLElementTagNameMap {
     'enchanted-textfield': EnchantedInputTextfield
   }
-}
-
+}